requirements.txt
python-telegram-bot==20.*
Flask==2.*
gunicorn==20.*